Revert "New mac machine aligning"
This reverts commit a220019c10fef9a7e5eec30fc4f63013957629bc.

+++ UIViewController+index.h
@@ -9,5 +9,5 @@
 #import <UIKit/UIKit.h>
 
 @interface UIViewController (index)
-@property (assign, nonatomic) int indexNumber;
+@property (assign, nonatomic) NSInteger indexNumber;
 @end
